Fixed Crash (and more)
Found a crash... Went to fix it. Ended up finding a fundamental issue with the way I wrote my code in the first place. Ended up spending a whole day on a problem that was super duper simple. Also I removed a comma & space from the comment variable on "patchEnemies". I am not sure why that was there in the first place, and I couldn't find any use of it, so I ended up removing it.